Move the Transform utility methods into a dedicated class.
